% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellContour_function.R
\name{CellContour}
\alias{CellContour}
\title{Determine Cell Contour}
\usage{
CellContour(x)
}
\arguments{
\item{x}{a 2D array indicating all pixels assigned to the cell shape.}
}
\description{
This function allows you to determine the cell contour based upon a black and white image array.
The function returns an array containing the number of contour points; x-position; y-position; total area.
}
\examples{
img <- readPNG("cell.png")
tmp<-array(CellContour(as.matrix(img)),dim=c(100000,4))
contour<-array(tmp,dim=c(tmp[1][1],4))
}
\keyword{contour}
