% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellContour_function.R
\name{CellContour}
\alias{CellContour}
\title{Determine Cell Contour}
\usage{
CellContour(x)
}
\arguments{
\item{array}{of x and y values of the image}
}
\description{
This function allows you to determine the cell contour based upon a black and white image
}
\examples{
CellContour()
}
\keyword{contour}
