% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOCOEFA_function.R
\name{LOCOEFA}
\alias{LOCOEFA}
\title{Calculate LOCO-EFA Coefficients}
\usage{
LOCOEFA(x)
}
\arguments{
\item{array}{of x and y coordinates of the outline}
}
\description{
This function allows you to calculate the LOCO-EFA coefficients based upon a cell outline
}
\examples{
LOCOEFA()
}
\keyword{LOCO-EFA}
